Move error catching to cypress support file. Add cypress to lint folders.
