+++ package.json
@@ -14,7 +14,7 @@
     "deploy": "firebase-ci deploy -s",
     "start": "cross-env NODE_ENV=development webpack-dashboard -p 1234 -c cyan -m -t barista -- node build/scripts/start",
     "start:simple": "cross-env NODE_ENV=development node build/scripts/start",
-    "lint": "eslint src functions server build project.config.js",
+    "lint": "eslint src functions server build cypress project.config.js",
     "lint:fix": "npm run lint -- --fix",
     "functions:start": "npm run start --prefix functions",
     "functions:build": "npm run build --prefix functions",
