+++ src/utils/components.js
@@ -1,8 +1,7 @@
-/* eslint-disable no-console */
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { pick, some, every } from 'lodash'
-import { isEmpty, isLoaded } from 'react-redux-firebase'
+import { pick, some } from 'lodash'
+import { isLoaded } from 'react-redux-firebase'
 import { mapProps, branch, renderComponent } from 'recompose'
 import LoadingSpinner from 'components/LoadingSpinner'
 
@@ -12,8 +11,9 @@ import LoadingSpinner from 'components/LoadingSpinner'
  * @param  {Function} condition - Condition function for when to show spinner
  * @return {HigherOrderComponent}
  */
-export const spinnerWhile = condition =>
-  branch(condition, renderComponent(LoadingSpinner))
+export function spinnerWhile(condition) {
+  return branch(condition, renderComponent(LoadingSpinner))
+}
 
 /**
  * Show a loading spinner while props are loading . Checks
@@ -36,12 +36,9 @@ export const spinnerWhile = condition =>
  * @param  {Array} propNames - List of prop names to check loading for
  * @return {HigherOrderComponent}
  */
-export const spinnerWhileLoading = propNames =>
-  spinnerWhile(
-    props =>
-      every(propNames, name => !isEmpty(props[name])) &&
-      some(propNames, name => !isLoaded(props[name]))
-  )
+export function spinnerWhileLoading(propNames) {
+  return spinnerWhile(props => some(propNames, name => !isLoaded(props[name])))
+}
 
 /**
  * HOC that logs props using console.log. Accepts an array list of prop names
@@ -62,37 +59,54 @@ export const spinnerWhileLoading = propNames =>
  * are logged
  * @return {HigherOrderComponent}
  */
-export const logProps = (propNames, logName = '') =>
-  mapProps(ownerProps => {
+export function logProps(propNames, logName = '') {
+  return mapProps(ownerProps => {
+    /* eslint-disable no-console */
     console.log(
       `${logName} props:`,
       propNames ? pick(ownerProps, propNames) : ownerProps
     )
+    /* eslint-enable no-console */
     return ownerProps
   })
-
-export const createWithFromContext = withVar => WrappedComponent => {
-  class WithFromContext extends Component {
-    render() {
-      const props = { [withVar]: this.context[withVar] }
-      if (this.context.store && this.context.store.dispatch) {
-        props.dispatch = this.context.store.dispatch
-      }
-      return <WrappedComponent {...this.props} {...props} />
-    }
-  }
-
-  WithFromContext.contextTypes = {
-    [withVar]: PropTypes.object.isRequired
-  }
-
-  return WithFromContext
 }
 
+/**
+ * Function which creates a HOC which gets a variable
+ * @param  {String} withVar - Name of variable to get from context
+ * @return {Component} React Higher Order Component
+ * @example Basic
+ * export const withRouter = createWithFromContext('router')
+ * // withRouter is an HOC that passes context.router as props.router
+ */
+export function createWithFromContext(withVar) {
+  return WrappedComponent => {
+    class WithFromContext extends Component {
+      render() {
+        const props = { [withVar]: this.context[withVar] }
+        if (this.context.store && this.context.store.dispatch) {
+          props.dispatch = this.context.store.dispatch
+        }
+        return <WrappedComponent {...this.props} {...props} />
+      }
+    }
+
+    WithFromContext.contextTypes = {
+      [withVar]: PropTypes.object.isRequired
+    }
+
+    return WithFromContext
+  }
+}
+
+/**
+ * HOC that adds router to props (from context)
+ * @return {HigherOrderComponent}
+ */
 export const withRouter = createWithFromContext('router')
 
 /**
- * HOC that adds store to props
+ * HOC that adds store to props (from context)
  * @return {HigherOrderComponent}
  */
 export const withStore = createWithFromContext('store')
